:package: improved structure
In this commit the structure is improved by removing some things from classes.py and putting it in different files plus also improving the new k!settings command a bit
