+++ killua/cogs/actions.py
@@ -5,11 +5,12 @@ import asyncio
 from typing import List, Union
 
 from killua.utils.checks import check
-from killua.utils.classes import Category, User, Button
-from killua.utils.paginator import View
+from killua.utils.classes import User
+from killua.static.enums import Category
+from killua.utils.interactions import View
 from killua.static.constants import ACTIONS
 
-class Select(discord.ui.Select):
+class SettingsSelect(discord.ui.Select):
     """Creates a select menu to change action settings"""
     def __init__(self, options, **kwargs):
         super().__init__(
@@ -22,6 +23,17 @@ class Select(discord.ui.Select):
         for opt in self.options:
             if opt.value in self.view.values:
                 opt.default = True
+
+class SettingsButton(discord.ui.Button):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    async def callback(self, interaction: discord.Interaction):
+        if not hasattr(self.view, "values"):
+            return await interaction.response.send_message("You have not changed any settings", ephemeral=True)
+        self.view.timed_out = False
+        self.view.stop()
 
 class Actions(commands.Cog):
 
@@ -192,6 +204,18 @@ class Actions(commands.Cog):
         """Snuggle up to a user and cuddle them with this command"""
         return await self.do_action(ctx, members)
 
+    def _get_view(self, id:int, current: dict) -> View:
+        options = [discord.SelectOption(label=k, value=k, default=v) for k, v in current.items()]
+        select = SettingsSelect(options, min_values=0, max_values=len(current))
+        button = SettingsButton(label="Save", style=discord.ButtonStyle.green, emoji="\U0001f4be")
+        view = View(user_id=id, timeout=100)
+        view.timed_out = True
+
+        view.add_item(select)
+        view.add_item(button)
+
+        return view
+
     @check()
     @commands.command(extras={"category": Category.ACTIONS}, usage="settings")
     async def settings(self, ctx):
@@ -213,41 +237,42 @@ class Actions(commands.Cog):
                 embed.add_field(name=action, value="✅", inline=False)
                 current[action] = True
         
-        options = [discord.SelectOption(label=k, value=k, default=v) for k, v in current.items()]
-        select = Select(options, min_values=0, max_values=len(current))
-        button = Button(label="Save", style=discord.ButtonStyle.green, emoji="\U0001f4be")
-        view = View(user_id=ctx.author.id, timeout=100)
-
-        view.add_item(select)
-        view.add_item(button)
+        view = self._get_view(ctx.author.id, current)
 
         msg = await ctx.send(embed=embed, view=view)
 
         await view.wait()
-        await view.disable(msg)
 
         if view.timed_out:
-            return
+            return await view.disable(msg)
 
         for val in view.values:
             current[val] = True
 
-        embed = discord.Embed.from_dict({
-            "title": "Settings",
-            "description": "By unticking a box users will no longer able to use that action on you",
-            "color": 0x1400ff
-        })
+        while True:
+            embed.clear_fields()
 
-        for action in ACTIONS.keys():
-            if action in view.values:
-                current[action] = True
-                embed.add_field(name=action, value="✅", inline=False)
-            else:
-                current[action] = False
-                embed.add_field(name=action, value="❌", inline=False)
+            for action in ACTIONS.keys():
+                if action in view.values:
+                    current[action] = True
+                    embed.add_field(name=action, value="✅", inline=False)
+                else:
+                    current[action] = False
+                    embed.add_field(name=action, value="❌", inline=False)
 
-        user.set_action_settings(current)
-        await msg.edit(embed=embed)
+            user.set_action_settings(current)
+            view = self._get_view(ctx.author.id, current)
+        
+            await msg.edit(embed=embed, view=view)
+
+            await view.wait()
+
+            if view.timed_out:
+                return await view.disable(msg)
+
+            for val in view.values:
+                current[val] = True
+
 
 Cog = Actions
