+++ killua/cogs/devstuff.py
@@ -4,7 +4,8 @@ from datetime import datetime, timedelta
 import re
 
 from killua.utils.checks import check
-from killua.utils.classes import User, Category #lgtm [py/unused-import]
+from killua.utils.classes import User #lgtm [py/unused-import]
+from killua.static.enums import Category
 from killua.static.cards import Card #lgtm [py/unused-import]
 from killua.static.constants import teams, guilds, blacklist, presence as pr, items, updates, UPDATE_CHANNEL #lgtm [py/unused-import]
