+++ killua/cogs/shop.py
@@ -8,10 +8,14 @@ from typing import Union, Tuple
 
 from killua.static.cards import Card
 from killua.static.constants import items, shop, FREE_SLOTS, ALLOWED_AMOUNT_MULTIPLE, PRICES, LOOTBOXES, editing
-from killua.utils.classes import Category, User, TodoList, PrintColors, CardNotFound, Button, ConfirmButton, CardLimitReached
+from killua.utils.classes import User, TodoList, CardNotFound, CardLimitReached
+from killua.static.enums import Category, PrintColors
+from killua.utils.interactions import Button, ConfirmButton
+
+
 from killua.utils.checks import check
-from killua.utils.paginator import DefaultEmbed, View, Paginator
-from killua.utils.help import Select
+from killua.utils.paginator import DefaultEmbed, Paginator
+from killua.utils.interactions import Select, View
 
 class ShopPaginator(Paginator):
     """A normal paginator with a button that returns to the original shop select menu"""
