+++ killua/utils/classes.py
@@ -41,17 +41,6 @@ class CheckFailure(Exception):
 class SuccessfullDefense(CheckFailure):
     pass
 
-class PrintColors:
-    HEADER = '\033[95m'
-    OKBLUE = '\033[94m'
-    OKCYAN = '\033[96m'
-    OKGREEN = '\033[92m'
-    WARNING = '\033[93m'
-    FAIL = '\033[91m'
-    ENDC = '\033[0m'
-    BOLD = '\033[1m'
-    UNDERLINE = '\033[4m'
-
 class PartialCard:
     """A class preventing a circular import by providing the bare minimum of methods and properties. Only used in this file"""
     def __init__(self, card_id:int):
@@ -235,129 +224,6 @@ class LootBox:
                 if user.is_entitled_to_double_jenny:
                     r *= 2
                 user.add_jenny(r)
-
-class Button(discord.ui.Button):
-
-    def __init__(self, **kwargs):
-        super().__init__(**kwargs)
-
-    async def callback(self, interaction: discord.Interaction):
-        self.view.value = self.custom_id
-        self.view.stop()
-
-class ConfirmButton(discord.ui.View):
-    """A button that is used to confirm a certain action or deny it"""
-    def __init__(self, user_id:int, **kwargs):
-        super().__init__(**kwargs)
-        self.user_id = user_id
-        self.timed_out = False # helps subclasses using Button to have set this to False
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if not (val := interaction.user.id == self.user_id):
-            await interaction.response.defer()
-        return val
-
-    async def disable(self, msg:discord.Message) -> discord.Message:
-        if len([c for c in self.children if not c.disabled]) == 0: # if every child is already disabled, we don't need to edit the message again
-            return
-
-        for child in self.children:
-            child.disabled = True
-
-        await msg.edit(view=self)
-
-    async def on_timeout(self):
-        self.value = False
-        self.timed_out = True
-    
-    @discord.ui.button(label="confirm", style=discord.ButtonStyle.green)
-    async def confirm(self, button: discord.ui.Button, interaction: discord.Interaction):
-        self.value = True
-        self.timed_out = False
-        self.stop()
-
-    @discord.ui.button(label="cancel", style=discord.ButtonStyle.red)
-    async def cancel(self, button: discord.ui.Button, interaction: discord.Interaction):
-        self.value = False
-        self.timed_out = False
-        self.stop()
-
-class Category(Enum):
-
-    ACTIONS = {
-        "name": "actions",
-        "description": "Commands that can be used to interact with other users, such as hugging them",
-        "emoji": {
-            "unicode": "\U0001f465",
-            "normal": ":busts_in_silhouette:"
-        }
-    }
-    CARDS = {
-        "name": "cards",
-        "description": "The greed island card system with monster, spell and item cards",
-        "emoji": {
-            "unicode": "<:card_number_46:811776158966218802>",
-            "normal": "<:card_number_46:811776158966218802>"
-        }
-    }
-    ECONOMY = {
-        "name": "economy",
-        "description": "Killua's economy with the currency Jenny",
-        "emoji": {
-            "unicode": "\U0001f3c6",
-            "normal": ":trophy:"
-        } 
-    }
-    MODERATION = {
-        "name": "moderation",
-        "description": "Moderation commands",
-        "emoji": {
-            "unicode": "\U0001f6e0",
-            "normal": ":tools:"
-        }
-    }
-    TODO = {
-        "name": "todo",
-        "description": "Todo lists on discord to help you be organised",
-        "emoji": {
-            "unicode": "\U0001f4dc",
-            "normal": ":scroll:"
-        }
-    }
-    FUN = {
-        "name": "fun",
-        "description": "Commands to play around with with friends to pass the time",
-        "emoji": {
-            "unicode": "\U0001f921",
-            "normal": ":clown:"
-        }
-    }
-    OTHER = {
-        "name": "other",
-        "description": "Commands that fit none of the other categories",
-        "emoji": {
-            "unicode": "<:killua_wink:769919176110112778>",
-            "normal": "<:killua_wink:769919176110112778>"
-        }
-    }
-
-    GAMES = {
-        "name": "games",
-        "description": "Games you can play with friends or alone",
-        "emoji": {
-            "unicode": "\U0001f3ae",
-            "normal": ":video_game:"
-        }
-    }
-
-    TAGS = {
-        "name": "tags",
-        "description": "Tags if you want to save some text. `[PREMIUM ONLY]`",
-        "emoji": {
-            "unicode": "\U0001f5c4",
-            "normal": ":file_cabinet:"
-        }
-    }
 
 # pillow logic contributed by DerUSBstick (Thank you!)
 class Book:
