+++ killua/utils/paginator.py
@@ -4,6 +4,7 @@ import discord
 from discord.ext import commands
 import datetime
 from inspect import iscoroutinefunction
+from .interactions import View
 
 from typing import List, Union, Type, TypeVar, Coroutine, Tuple, Callable
 
@@ -32,32 +33,6 @@ class DefaultEmbed(discord.Embed):
         self.color = 0x1400ff
         self.timestamp = datetime.datetime.utcnow()
         self.set_footer()
-
-class View(discord.ui.View):
-    """Subclassing this for buttons enabled us to not have to define interaction_check anymore, also not if we want a select menu"""
-    def __init__(self, user_id:int, **kwargs):
-        super().__init__(**kwargs)
-        self.user_id = user_id
-        self.value = None
-        self.timed_out = False
-
-    async def on_timeout(self):
-        self.timed_out = True
-
-    async def interaction_check(self, interaction: discord.Interaction) -> bool:
-        if not (val := interaction.user.id == self.user_id):
-            await interaction.response.defer()
-        return val
-
-    async def disable(self, msg:discord.Message) -> Union[discord.Message, None]:
-        """"Disables the children inside of the view"""
-        if not [c for c in self.children if not c.disabled]: # if every child is already disabled, we don't need to edit the message again
-            return
-
-        for c in self.children:
-            c.disabled = True
-
-        await msg.edit(view=self)
 
 class Buttons(View):
     """The core of the paginator"""
